requirements.txt
coloredlogs==15.0.1
gpiozero==2.0.1
mutagen==1.47.0
numpy==1.24.2
Pillow==11.1.0
pygame==2.1.2
smbus==1.1.post2
spidev==3.5
